scripts/list_models.py
import os

import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv(".env")

api_key = os.environ.get("GOOGLE_API_KEY")
if not api_key:
    print("No API Key found")
    exit(1)

genai.configure(api_key=api_key)

print("Available Embedding Models:")
try:
    for m in genai.list_models():
        if "embedContent" in m.supported_generation_methods:
            print(m.name)
except Exception as e:
    print(f"Error listing models: {e}")
