requirements.txt
pydantic
fastapi
uvicorn
python-dotenv
mcp
langchain
langgraph
langchain-google-genai
langchain-openai
langchain-anthropic
xai-sdk
structlog
langgraph-checkpoint-sqlite
pyyaml
duckduckgo-search<7.0.0
langchain-community
chromadb
langchain-chroma
pytest
pytest-asyncio
pytest-cov
